[chip-test] Add csrng_kat_test
Adds basic known answer test based on NIST CAVP test vector. This commit
also simplifies the csrng smoketest, and enables it on FPGA and
verilator test platforms.
